% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{dosing_options}
\alias{dosing_options}
\alias{dosing_setOption}
\alias{dosing_getOption}
\title{Options for the donkey dosing model}
\usage{
dosing_setOption(...)

dosing_getOption(name)

dosing_options()
}
\arguments{
\item{...}{a named list of options to set}

\item{name}{the name of the option to retrieve}
}
\description{
These functions allow global options to be get and set
}
