% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_weather.R
\name{extract_weather}
\alias{extract_weather}
\title{Read weather from a local cache and format it for use with the Donkey Dosing model}
\usage{
extract_weather(
  path,
  year = format(Sys.Date(), "\%Y"),
  use_cache = TRUE,
  impute_missing = FALSE,
  include_date = FALSE,
  type = "OpenWeather"
)
}
\arguments{
\item{path}{a path to a folder containing the raw weather data}

\item{year}{the year for which to extract data}

\item{use_cache}{should the cache file be used where possible?}

\item{impute_missing}{should missing data be imputed?}

\item{include_date}{should the date be included in the return value?}

\item{type}{the source type for weather data (only OpenWeather is currently permitted)}
}
\description{
Read weather from a local cache and format it for use with the Donkey Dosing model
}
