% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_openweather.R
\name{scrape_openweather}
\alias{scrape_openweather}
\title{Scrape OpenWeather data for the Donkey Dosing model}
\usage{
scrape_openweather(path, api_key = NULL, lat = 50.725, lon = -3.175)
}
\arguments{
\item{path}{a path to a folder containing the raw weather data}

\item{api_key}{either an OpenWeather API key, or NULL in which case the key will be read from an "OpenWeatherKey.txt" file in the specified path}

\item{lat}{the latitude for the location used for scraping (default is a point roughly in the middle of the Donkey Sanctuary):}

\item{lon}{the longitude for the location used for scraping (default is a point roughly in the middle of the Donkey Sanctuary):}
}
\value{
a character vector of paths to which weather observations have been saved, invisibly
}
\description{
Scrape OpenWeather data for the Donkey Dosing model
}
