% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScrapeWeather.R
\name{ScrapeWeather}
\alias{ScrapeWeather}
\title{Scrape the requrired weather data}
\usage{
ScrapeWeather(input_file = NA, output_file = NA, ...)
}
\arguments{
\item{input_file}{either a path to an Excel or CSV file containing previously recorded (or scraped) weather data, a data frame of the same, or NA}

\item{output_file}{either a path to an Excel or CSV file to which the combined data will be written, or NA}

\item{...}{additional arguments to be passed through to the darksky package}
}
\value{
A data frame with the available weather data in an appropriate format for the donkey dosing model
}
\description{
This function can be used to scrape weather data from a website (intended to be run automatically on a regular basis).  If historical data is provided as input then the scraped data will be combined with this.
}
